Isolated pre-processor functions
This commit has moved over all the pre processor functions from
the `instruction.c` file. Since the `instruction.c` file is already
fairly crowded from all the instruction encoder tables (With more
comming soon), there was no reason for the pre-processor functions
to be crowding around with no direct benefits to the tables *iteself*

Therefore, these pre-processors has been ported over to a new file
and included *directly* (Which I know is reletivly bad practice)
